Flask/requirements.txt
Flask==1.1.1
opencv-python-headless
numpy
jsonpickle
google-cloud-storage
keras==2.2.4
tensorflow==1.13.1
gunicorn==19.7.1
pillow